Separated each classes into individual java file
